rami-app/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-groq
langchain-community
chromadb
sentence-transformers
jinja2
pydantic
markdown2
hf_xet
